renaming files to kebab case
renamed testing
